+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com) and this p
 
 ## [Unreleased](https://github.com/Nutzzz/GameCollector/compare/v4.2.0.5...HEAD)
 
+- Ignore "Audience" entries for Epic
 - Set single-file publish options for GameFinder.Example, and rename its assembly to output GameCollector.exe
 
 **TODO**: